requirements.txt
pip
fastapi
filelock
fire
gradio
scipy
numpy
Pillow
pydantic
Requests
torch
torchvision
sentencepiece
diffusers
transformers
uvicorn
einops
timm
bitsandbytes
accelerate
spaces
hf_transfer
